refactor: view scripts
